Feat(data): Add Formations To Human Fleets (#209)
This adds the "Vic" formation to the vast majority of civilian and authority ships in human space; with the intent on having defined fleets actually look a little bit more like actual fleets. The impact is minimal to do problems with the fleet code causing fleets to not stick together and prioritize landing and hyperjumping over forming and maintaining a fleet, but it sometimes happens, and when it does, it looks cool.