feat(data): Adding slots to the Avgi (#219)
This adds thruster, steering, lateral, and reverse slots to the various Avgi ships; as well as adding the requirements therefor to the various thrusters. Worth noting that a couple of these specifically denote that the drive gimbals to provide "full six axis control", and as such has had lateral thrust included. If something provides thrust, reverse thrust, and steering... It seems appropriate that it would handle lateral thrust too.

Co-authored-by: Loymdayddaud <145969603+TheGiraffe3@users.noreply.github.com>

To-do: Add lateral flares to the Avgi ships. They already have tons of flare points listed for thrust, reverse thrust, and steering; so it should be just a matter of cut and paste and figuring out which to use; but there certainly are a *lot* of flare points...